fix: handle text msg recall fix: BaseBotException is not catched by nonebot (and printed) fix: database startup at module level to enable load a single plugin
